パディングできたはず後々悪さしそうな設定にしているためバグったらその辺に注意

+++ Engine/Game/Light/PointLight.h
@@ -22,7 +22,7 @@ public:
 		//‹——£Œ¸ŠŒW”
 		Vector3 lightAtten;
 
-		uint16_t active;
+		float active;
 
 	};
